requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.7.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
dydantic==0.0.8
ecdsa==0.19.0
fastapi==0.108.0
filelock==3.20.0
frozenlist==1.5.0
fsspec==2025.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
idna==3.10
Jinja2==3.1.6
jiter==0.8.2
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.22
langchain-community==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.8
langchain-text-splitters==0.3.11
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-checkpoint-postgres==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langmem==0.0.30
langsmith==0.3.45
MarkupSafe==3.0.3
marshmallow==3.23.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.5
numpy==1.26.4
openai==1.58.1
orjson==3.10.13
ormsgpack==1.11.0
packaging==24.2
passlib==1.7.4
pillow==12.0.0
propcache==0.2.1
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pypdf==5.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==3.3.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.32.0.post1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.21.4
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
trustcall==0.0.39
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.18.3
zstandard==0.23.0
